refactor: package declaration
refactored package declaration of the test and main files within the upload
