+++ upload-service/src/main/java/com/App.java
@@ -1,4 +1,4 @@
-package upload;
+package com;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
