+++ upload-service/src/main/java/com/util/ImageUtils.java
@@ -1,4 +1,4 @@
-package upload.util;
+package com.util;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
